+++ app/models/facilities_management/buyer_detail.rb
@@ -2,25 +2,39 @@ module FacilitiesManagement
   class BuyerDetail < ApplicationRecord
     belongs_to :user, inverse_of: :buyer_detail
 
-    attr_accessor :postcode_entry
-
-    after_initialize do |bd|
-      bd.postcode_entry = bd.organisation_address_postcode if bd.postcode_entry.blank?
-    end
-
     MAX_FIELD_LENGTH = 255
     SECTORS = %w[culture_media_and_sport defence_and_security education government_policy health infrastructure local_community_and_housing].freeze
 
-    validates :full_name, presence: true, length: { maximum: MAX_FIELD_LENGTH }, on: :update
-    validates :job_title, presence: true, length: { maximum: MAX_FIELD_LENGTH }, on: :update
-    validates :telephone_number, presence: true, format: { with: /\A[\s()\d-]{10,14}\d\z/ }, length: { maximum: 15 }, on: :update
-    validates :telephone_number, numericality: { greater_than: 0, message: :blank }, on: :update
-    validates :organisation_name, length: { maximum: MAX_FIELD_LENGTH }, presence: true, on: :update
+    with_options on: %i[update personal_details] do
+      validates :full_name, presence: true, length: { maximum: MAX_FIELD_LENGTH }
 
-    include AddressValidator
+      validates :job_title, presence: true, length: { maximum: MAX_FIELD_LENGTH }
 
-    validates :sector, inclusion: { in: SECTORS }, on: :update
-    validates :contact_opt_in, inclusion: { in: [true, false] }, on: :update
+      validates :telephone_number, presence: true, format: { with: /\A[\s()\d-]{10,14}\d\z/ }, length: { maximum: 15 }
+      validates :telephone_number, numericality: { greater_than: 0, message: :blank }
+    end
+
+    with_options on: %i[update organisation_details] do
+      validates :organisation_name, length: { maximum: MAX_FIELD_LENGTH }, presence: true
+
+      validates :organisation_address_line_1, length: { maximum: MAX_FIELD_LENGTH }, presence: true
+
+      validates :organisation_address_line_2, length: { maximum: MAX_FIELD_LENGTH }
+
+      validates :organisation_address_town, length: { maximum: MAX_FIELD_LENGTH }, presence: true
+
+      validates :organisation_address_county, length: { maximum: MAX_FIELD_LENGTH }
+
+      validates :organisation_address_postcode, presence: true
+
+      validate :postcode_format, if: -> { organisation_address_postcode.present? }
+
+      validates :sector, inclusion: { in: SECTORS }
+    end
+
+    with_options on: %i[update contact_preferences] do
+      validates :contact_opt_in, inclusion: { in: [true, false] }
+    end
 
     delegate :email, to: :user
 
@@ -29,7 +43,19 @@ module FacilitiesManagement
     end
 
     def sector_name
-      sector.nil? ? I18n.t("facilities_management.buyer_details.edit.central_government.options.#{central_government}") : I18n.t("facilities_management.buyer_details.edit.sector.options.#{sector}")
+      sector.nil? ? I18n.t("facilities_management.buyer_details.sections.organisation_details.sector.options.#{central_government}") : I18n.t("facilities_management.buyer_details.sections.organisation_details.sector.options.#{sector}")
+    end
+
+    def details_complete?
+      valid?(:update)
+    end
+
+    private
+
+    def postcode_format
+      postcode = UKPostcode.parse(organisation_address_postcode)
+
+      postcode.full_valid? ? errors.delete(:organisation_address_postcode) : errors.add(:organisation_address_postcode, :invalid)
     end
   end
 end