+++ spec/controllers/facilities_management/buyer_details_controller_spec.rb
@@ -1,34 +1,39 @@
 require 'rails_helper'
-# rubocop:disable RSpec/NamedSubject
+
 module FacilitiesManagement
   RSpec.describe BuyerDetailsController do
     let(:default_params) { { service: 'facilities_management', framework: framework } }
-    let(:framework) { 'RM6232' }
+    let(:framework) { 'RM6378' }
+    let(:user) { controller.current_user }
 
-    include_context 'and RM6232 is live'
-
-    render_views
-
-    describe 'GET edit with buyer' do
+    describe 'GET show' do
       context 'when logged in with buyer details' do
         login_fm_buyer_with_details
 
-        before { get :edit, params: { id: subject.current_user.id } }
+        before { get :show, params: { id: user.id } }
 
-        it 'renders edit template' do
+        it 'renders show template' do
           expect(response).to have_http_status(:ok)
-          expect(response).to render_template(:edit)
+          expect(response).to render_template(:show)
+        end
+
+        it 'sets buyer_detail' do
+          expect(assigns(:buyer_detail)).to be_present
         end
       end
 
       context 'when logged in without buyer details' do
         login_fm_buyer
 
-        before { get :edit, params: { id: subject.current_user.id } }
+        before { get :show, params: { id: user.id } }
 
-        it 'renders edit template' do
+        it 'renders show template' do
           expect(response).to have_http_status(:ok)
-          expect(response).to render_template(:edit)
+          expect(response).to render_template(:show)
+        end
+
+        it 'sets buyer_detail' do
+          expect(assigns(:buyer_detail)).to be_present
         end
       end
 
@@ -37,7 +42,7 @@ module FacilitiesManagement
 
         let(:framework) { 'RM3840' }
 
-        before { get :edit, params: { id: subject.current_user.id } }
+        before { get :show, params: { id: user.id } }
 
         it 'renders the unrecognised framework page with the right http status' do
           expect(response).to render_template('home/unrecognised_framework')
@@ -51,26 +56,54 @@ module FacilitiesManagement
       end
     end
 
-    describe 'GET edit_address' do
-      context 'when logged in with buyer details' do
-        login_fm_buyer_with_details
+    describe 'GET edit' do
+      login_fm_buyer_with_details
 
-        before { get :edit_address, params: { buyer_detail_id: subject.current_user.id } }
+      before { get :edit, params: { id: user.buyer_detail.id, section: section } }
 
-        it 'renders edit template' do
-          expect(response).to have_http_status(:ok)
-          expect(response).to render_template(:edit_address)
+      shared_examples 'when testing a section' do
+        it 'sets buyer_detail' do
+          expect(assigns(:buyer_detail)).to be_present
+        end
+
+        it 'sets the back_path' do
+          expect(assigns(:back_path)).to eq("/facilities-management/RM6378/buyer-details/#{user.buyer_detail.id}")
+          expect(assigns(:back_text)).to eq('Your details')
+        end
+
+        context 'when considering the templates' do
+          render_views
+
+          it 'renders section partial template' do
+            expect(response).to have_http_status(:ok)
+            expect(response).to render_template(partial: "facilities_management/buyer_details/sections/_#{section}")
+          end
         end
       end
 
-      context 'when logged in without buyer details' do
-        login_fm_buyer
+      context 'when the section is personal_details' do
+        let(:section) { :personal_details }
 
-        before { get :edit_address, params: { buyer_detail_id: subject.current_user.id } }
+        include_context 'when testing a section'
+      end
 
-        it 'renders edit template' do
-          expect(response).to have_http_status(:ok)
-          expect(response).to render_template(:edit_address)
+      context 'when the section is organisation_details' do
+        let(:section) { :organisation_details }
+
+        include_context 'when testing a section'
+      end
+
+      context 'when the section is contact_preferences' do
+        let(:section) { :contact_preferences }
+
+        include_context 'when testing a section'
+      end
+
+      context 'when the section is unexpected' do
+        let(:section) { :something_else }
+
+        it 'redirects to the show page' do
+          expect(response).to redirect_to("/facilities-management/RM6378/buyer-details/#{user.buyer_detail.id}")
         end
       end
     end
@@ -78,58 +111,80 @@ module FacilitiesManagement
     describe 'PUT update' do
       login_fm_buyer_with_details
 
-      context 'when updating from the edit page' do
-        before { put :update, params: { id: subject.current_user.id, facilities_management_buyer_detail: { full_name: } } }
+      before { put :update, params: { id: user.buyer_detail.id, section: section, facilities_management_buyer_detail: facilities_management_buyer_detail } }
 
-        context 'when there are errors' do
-          let(:full_name) { '' }
+      shared_examples 'when testing a section' do
+        it 'sets buyer_detail' do
+          expect(assigns(:buyer_detail)).to be_present
+        end
 
-          it 'renders the edit template' do
-            expect(response).to render_template(:edit)
+        it 'sets the back_path' do
+          expect(assigns(:back_path)).to eq("/facilities-management/RM6378/buyer-details/#{user.buyer_detail.id}")
+          expect(assigns(:back_text)).to eq('Your details')
+        end
+
+        context 'when it is valid' do
+          it 'redirects to the show page' do
+            expect(response).to redirect_to("/facilities-management/RM6378/buyer-details/#{user.buyer_detail.id}")
+          end
+
+          it 'updates the buyer details' do
+            expect(user.reload.buyer_detail.attributes.deep_symbolize_keys.slice(*facilities_management_buyer_detail.keys)).to eq(facilities_management_buyer_detail)
           end
         end
 
-        context 'when there are no errors' do
-          let(:full_name) { 'Fred Flintstone' }
+        context 'when it is invalid' do
+          let(:facilities_management_buyer_detail) { facilities_management_buyer_detail_invalid }
 
-          it 'redirects to facilities_management_index_path' do
-            expect(response).to redirect_to facilities_management_index_path(framework:)
+          render_views
+
+          it 'has errors on the buyer detail' do
+            expect(assigns(:buyer_detail).errors).to be_present
           end
 
-          it 'updates the buyer name' do
-            subject.current_user.reload
-
-            expect(subject.current_user.buyer_detail.full_name).to eq full_name
+          it 'renders section partial template' do
+            expect(response).to have_http_status(:ok)
+            expect(response).to render_template(partial: "facilities_management/buyer_details/sections/_#{section}")
           end
         end
       end
 
-      context 'when updating from the edit_address page' do
-        before { put :update, params: { id: subject.current_user.id, context: :update_address, facilities_management_buyer_detail: { organisation_address_line_1: } } }
+      context 'when the section is personal_details' do
+        let(:section) { :personal_details }
 
-        context 'when there are errors' do
-          let(:organisation_address_line_1) { '' }
+        let(:facilities_management_buyer_detail) { { full_name: 'Zote the Mighty', job_title: 'Knight', telephone_number: '07123456789' } }
+        let(:facilities_management_buyer_detail_invalid) { { full_name: '', job_title: '', telephone_number: '' } }
 
-          it 'renders the edit_address template' do
-            expect(response).to render_template(:edit_address)
-          end
-        end
+        include_context 'when testing a section'
+      end
 
-        context 'when there are no errors' do
-          let(:organisation_address_line_1) { '9 Downing Street' }
+      context 'when the section is organisation_details' do
+        let(:section) { :organisation_details }
 
-          it 'redirects to edit_facilities_management_buyer_detail_path' do
-            expect(response).to redirect_to edit_facilities_management_buyer_detail_path(framework, subject.current_user)
-          end
+        let(:facilities_management_buyer_detail) { { organisation_name: 'The Knight', organisation_address_line_1: 'Hollow Nest', organisation_address_line_2: '', organisation_address_town: 'The City', organisation_address_county: '', organisation_address_postcode: 'ST13 4AA', sector: 'defence_and_security' } }
+        let(:facilities_management_buyer_detail_invalid) { { organisation_name: '', organisation_address_line_1: '', organisation_address_line_2: '', organisation_address_town: '', organisation_address_county: '', organisation_address_postcode: '', sector: '' } }
 
-          it 'updates the buyer name' do
-            subject.current_user.reload
+        include_context 'when testing a section'
+      end
 
-            expect(subject.current_user.buyer_detail.organisation_address_line_1).to eq organisation_address_line_1
-          end
+      context 'when the section is contact_preferences' do
+        let(:section) { :contact_preferences }
+
+        let(:facilities_management_buyer_detail) { { contact_opt_in: true } }
+        let(:facilities_management_buyer_detail_invalid) { { contact_opt_in: '' } }
+
+        include_context 'when testing a section'
+      end
+
+      context 'when the section is unexpected' do
+        let(:section) { :something_else }
+
+        let(:facilities_management_buyer_detail) { { full_name: 'Zote the Mighty', job_title: 'Knight', telephone_number: '07123456789' } }
+
+        it 'redirects to the show page' do
+          expect(response).to redirect_to("/facilities-management/RM6378/buyer-details/#{user.buyer_detail.id}")
         end
       end
     end
   end
 end
-# rubocop:enable RSpec/NamedSubject