+++ spec/models/facilities_management/buyer_detail_spec.rb
@@ -126,11 +126,10 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
 
         it 'is invalid' do
           expect(buyer_detail.valid?(:update)).to be false
-          expect(buyer_detail.valid?(:update_address)).to be false
         end
 
         it 'has the correct error message' do
-          buyer_detail.valid?(:update_address)
+          buyer_detail.valid?(:update)
 
           expect(buyer_detail.errors[:organisation_address_postcode].first).to eq 'Enter a valid postcode, for example SW1A 1AA'
         end
@@ -141,11 +140,10 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
 
         it 'is invalid' do
           expect(buyer_detail.valid?(:update)).to be false
-          expect(buyer_detail.valid?(:update_address)).to be false
         end
 
         it 'has the correct error message' do
-          buyer_detail.valid?(:update_address)
+          buyer_detail.valid?(:update)
 
           expect(buyer_detail.errors[:organisation_address_postcode].first).to eq 'Enter a valid postcode, for example SW1A 1AA'
         end
@@ -159,7 +157,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_line_1) { nil }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
+          expect(buyer_detail.valid?(:update)).to be false
         end
       end
 
@@ -167,7 +165,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_line_1) { 'a' * 256 }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
+          expect(buyer_detail.valid?(:update)).to be false
         end
       end
     end
@@ -179,7 +177,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_line_2) { nil }
 
         it 'is valid' do
-          expect(buyer_detail.valid?(:update_address)).to be true
+          expect(buyer_detail.valid?(:update)).to be true
         end
       end
 
@@ -187,7 +185,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_line_2) { 'a' * 256 }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
+          expect(buyer_detail.valid?(:update)).to be false
         end
       end
     end
@@ -199,7 +197,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_town) { nil }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
+          expect(buyer_detail.valid?(:update)).to be false
         end
       end
 
@@ -207,7 +205,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_town) { 'a' * 256 }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
+          expect(buyer_detail.valid?(:update)).to be false
         end
       end
     end
@@ -219,7 +217,7 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_county) { nil }
 
         it 'is valid' do
-          expect(buyer_detail.valid?(:update_address)).to be true
+          expect(buyer_detail.valid?(:update)).to be true
         end
       end
 
@@ -227,81 +225,10 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
         let(:organisation_address_county) { 'a' * 256 }
 
         it 'is invalid' do
-          expect(buyer_detail.valid?(:update_address)).to be false
-        end
-      end
-    end
-
-    # rubocop:disable RSpec/NestedGroups
-    context 'when considering address selection' do
-      let(:organisation_address_postcode) { 'ST161AA' }
-      let(:organisation_address_line_1) { 'The Globe Theatre' }
-      let(:organisation_address_town) { 'London Town' }
-
-      before { buyer_detail.assign_attributes(organisation_address_postcode:, organisation_address_line_1:, organisation_address_town:) }
-
-      context 'when the postcode is not valid' do
-        let(:organisation_address_postcode) { '' }
-
-        it 'is not valid' do
           expect(buyer_detail.valid?(:update)).to be false
         end
-
-        it 'does not have an error message on address selection' do
-          buyer_detail.valid?(:update)
-          expect(buyer_detail.errors[:base].any?).to be false
-        end
-      end
-
-      context 'when the postocde is valid' do
-        context 'and address_line_1 is missing' do
-          let(:organisation_address_line_1) { '' }
-
-          it 'is not valid' do
-            expect(buyer_detail.valid?(:update)).to be false
-          end
-
-          it 'has the correct error message' do
-            buyer_detail.valid?(:update)
-            expect(buyer_detail.errors[:base].first).to eq 'You must select an address to save your details'
-          end
-        end
-
-        context 'and address_town is missing' do
-          let(:organisation_address_town) { '' }
-
-          it 'is not valid' do
-            expect(buyer_detail.valid?(:update)).to be false
-          end
-
-          it 'has the correct error message' do
-            buyer_detail.valid?(:update)
-            expect(buyer_detail.errors[:base].first).to eq 'You must select an address to save your details'
-          end
-        end
-
-        context 'and address_line_1 and address_town are missing' do
-          let(:organisation_address_line_1) { '' }
-          let(:organisation_address_town) { '' }
-
-          it 'is not valid' do
-            expect(buyer_detail.valid?(:update)).to be false
-          end
-
-          it 'has the correct error message' do
-            buyer_detail.valid?(:update)
-            expect(buyer_detail.errors[:base].first).to eq 'You must select an address to save your details'
-          end
-        end
-      end
-
-      context 'when all parts are valid' do
-        it 'is valid' do
-          expect(buyer_detail.valid?(:update)).to be true
-        end
       end
     end
-    # rubocop:enable RSpec/NestedGroups
 
     context 'when considering sector' do
       before { buyer_detail.sector = sector }
@@ -473,6 +400,22 @@ RSpec.describe FacilitiesManagement::BuyerDetail do
 
       it 'returns Culture, Media and Sport' do
         expect(result).to eq 'Culture, Media and Sport'
+      end
+    end
+  end
+
+  describe '#details_complete?' do
+    context 'when all attributes are there' do
+      it 'returns true' do
+        expect(buyer_detail.details_complete?).to be(true)
+      end
+    end
+
+    context 'when some attributes are missing' do
+      before { buyer_detail.full_name = nil }
+
+      it 'returns true' do
+        expect(buyer_detail.details_complete?).to be(false)
       end
     end
   end